#25 - refactor: DTO 디렉토리 변경
