+++ src/main/java/cotato/bookitlist/auth/dto/request/ReissueRequest.java
@@ -1,4 +1,4 @@
-package cotato.bookitlist.auth.dto;
+package cotato.bookitlist.auth.dto.request;
 
 public record ReissueRequest(
         String refreshToken
