+++ src/main/java/cotato/bookitlist/config/security/jwt/JwtTokenProvider.java
@@ -2,15 +2,17 @@ package cotato.bookitlist.config.security.jwt;
 
 import cotato.bookitlist.config.security.jwt.dto.AccessTokenInfo;
 import cotato.bookitlist.config.security.jwt.properties.JwtProperties;
-import io.jsonwebtoken.*;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
 
 import java.nio.charset.StandardCharsets;
 import java.security.Key;
 import java.util.Date;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
 
 @RequiredArgsConstructor
 @Component
